大量模块使用gorm替换xorm

+++ main/app/system/model/monitor/logininfor/logininfor.go
@@ -62,7 +62,7 @@ func SelectPageList(param *SelectPageReq) (*[]Entity, *lv_web.Paging, error) {
 		return nil, p, errors.New("读取行数失败")
 	}
 
-	p = lv_web.CreatePaging(param.PageNum, param.PageSize, int(total))
+	p = lv_web.CreatePaging(param.PageNum, param.PageSize, int64(total))
 
 	if param.OrderByColumn != "" {
 		model.OrderBy(param.OrderByColumn + " " + param.IsAsc + " ")
